Refactor: Split classes into separate files
- Move DatabaseManager to packages/server/src/db/DatabaseManager.js
- Move BaseRepository to packages/server/src/db/BaseRepository.js
- Move ProjectRepository to packages/server/src/db/ProjectRepository.js
- Move SessionRepository to packages/server/src/db/SessionRepository.js
- Move MessageRepository to packages/server/src/db/MessageRepository.js
- Move CanvasItemRepository to packages/server/src/db/CanvasItemRepository.js
- Move SessionNoteRepository to packages/server/src/db/SessionNoteRepository.js
- Move WebSocketManager to packages/server/src/ws/WebSocketManager.js
- Move ApiClient to packages/web/src/api/ApiClient.js
- Add barrel exports (index.js) for each module
- Keep original files as re-exports for backward compatibility
