+++ packages/server/src/db/DatabaseManager.js
@@ -1,0 +1,75 @@
+import Database from 'better-sqlite3';
+import { readFileSync } from 'fs';
+import { fileURLToPath } from 'url';
+import { dirname, join } from 'path';
+import crypto from 'crypto';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+/**
+ * Database manager class - singleton pattern for managing SQLite connection
+ */
+export class DatabaseManager {
+  #db = null;
+
+  /**
+   * Initialize database with WAL mode and foreign keys
+   * @param {string} dbPath - Path to database file (use ':memory:' for in-memory)
+   * @returns {Database.Database}
+   */
+  init(dbPath = 'claudetools.db') {
+    this.#db = new Database(dbPath);
+
+    // Enable WAL mode and foreign keys
+    this.#db.pragma('journal_mode = WAL');
+    this.#db.pragma('foreign_keys = ON');
+
+    // Run schema
+    const schema = readFileSync(join(__dirname, '..', 'schema.sql'), 'utf-8');
+    this.#db.exec(schema);
+
+    return this.#db;
+  }
+
+  /**
+   * Get the database instance
+   * @returns {Database.Database}
+   */
+  get() {
+    if (!this.#db) {
+      throw new Error('Database not initialized. Call init first.');
+    }
+    return this.#db;
+  }
+
+  /**
+   * Close the database connection
+   */
+  close() {
+    if (this.#db) {
+      this.#db.close();
+      this.#db = null;
+    }
+  }
+
+  /**
+   * Generate a UUID
+   * @returns {string}
+   */
+  generateId() {
+    return crypto.randomUUID();
+  }
+
+  /**
+   * Run a function inside a transaction
+   * @param {Function} fn - Function to run in transaction
+   * @returns {any}
+   */
+  transaction(fn) {
+    return this.get().transaction(fn)();
+  }
+}
+
+// Singleton instance
+export const databaseManager = new DatabaseManager();
