+++ packages/server/src/db/SessionRepository.js
@@ -1,0 +1,95 @@
+import { BaseRepository } from './BaseRepository.js';
+import { databaseManager } from './DatabaseManager.js';
+import { messages } from './index.js';
+
+/**
+ * Session repository class
+ */
+export class SessionRepository extends BaseRepository {
+  constructor() {
+    super('sessions', SessionRepository.#mapSession);
+  }
+
+  static #mapSession(row) {
+    return {
+      id: row.id,
+      projectId: row.project_id,
+      name: row.name,
+      status: row.status,
+      mode: row.mode,
+      gitBranch: row.git_branch,
+      gitWorktree: row.git_worktree,
+      prUrl: row.pr_url,
+      error: row.error,
+      createdAt: row.created_at,
+      updatedAt: row.updated_at,
+    };
+  }
+
+  create(projectId, name, prompt, mode = 'standard', gitBranch = null) {
+    const id = databaseManager.generateId();
+    const now = Date.now();
+    this.db
+      .prepare(
+        `INSERT INTO sessions (id, project_id, name, status, mode, git_branch, created_at, updated_at)
+         VALUES (?, ?, ?, 'starting', ?, ?, ?, ?)`
+      )
+      .run(id, projectId, name, mode, gitBranch, now, now);
+
+    // Create initial user message
+    messages.create(id, 'user', prompt);
+
+    return this.getById(id);
+  }
+
+  getByProjectId(projectId) {
+    const rows = this.db
+      .prepare('SELECT * FROM sessions WHERE project_id = ? ORDER BY updated_at DESC')
+      .all(projectId);
+    return this.mapAll(rows);
+  }
+
+  update(id, data) {
+    const updates = [];
+    const values = [];
+
+    if (data.name !== undefined) {
+      updates.push('name = ?');
+      values.push(data.name);
+    }
+    if (data.status !== undefined) {
+      updates.push('status = ?');
+      values.push(data.status);
+    }
+    if (data.mode !== undefined) {
+      updates.push('mode = ?');
+      values.push(data.mode);
+    }
+    if (data.gitBranch !== undefined) {
+      updates.push('git_branch = ?');
+      values.push(data.gitBranch);
+    }
+    if (data.gitWorktree !== undefined) {
+      updates.push('git_worktree = ?');
+      values.push(data.gitWorktree);
+    }
+    if (data.prUrl !== undefined) {
+      updates.push('pr_url = ?');
+      values.push(data.prUrl);
+    }
+    if (data.error !== undefined) {
+      updates.push('error = ?');
+      values.push(data.error);
+    }
+
+    if (updates.length === 0) return this.getById(id);
+
+    updates.push('updated_at = ?');
+    values.push(Date.now());
+    values.push(id);
+
+    this.db.prepare(`UPDATE sessions SET ${updates.join(', ')} WHERE id = ?`).run(...values);
+
+    return this.getById(id);
+  }
+}
