+++ packages/server/src/ws/index.js
@@ -1,0 +1,40 @@
+export { WebSocketManager, webSocketManager } from './WebSocketManager.js';
+
+// Legacy function exports for backward compatibility
+import { webSocketManager } from './WebSocketManager.js';
+
+/**
+ * Initialize WebSocket server
+ * @param {import('http').Server} server
+ * @returns {import('ws').WebSocketServer}
+ */
+export function initWebSocket(server) {
+  return webSocketManager.init(server);
+}
+
+/**
+ * Broadcast message to all connected clients
+ * @param {string} type
+ * @param {Object} payload
+ */
+export function broadcast(type, payload) {
+  webSocketManager.broadcast(type, payload);
+}
+
+/**
+ * Broadcast message to clients subscribed to a session
+ * @param {string} sessionId
+ * @param {string} type
+ * @param {Object} payload
+ */
+export function broadcastToSession(sessionId, type, payload) {
+  webSocketManager.broadcastToSession(sessionId, type, payload);
+}
+
+/**
+ * Get WebSocket server instance
+ * @returns {import('ws').WebSocketServer|null}
+ */
+export function getWebSocketServer() {
+  return webSocketManager.getServer();
+}
