+++ packages/web/src/api/index.js
@@ -1,0 +1,1 @@
+export { ApiClient, api } from './ApiClient.js';
